% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_destin.R
\name{create_destin}
\alias{create_destin}
\title{Creates a directory \code{destin}, if it doesn't already exist.}
\usage{
create_destin(destin)
}
\arguments{
\item{destin}{Destination directory.}
}
\description{
Creates a directory \code{destin}, if it doesn't already exist.
}
