% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_args.R
\name{parse_args}
\alias{parse_args}
\title{Parse command line arguments and assigns the values of them. |args| is meant
to just be additional command line arguments.}
\usage{
parse_args(args, verbose = FALSE)
}
\arguments{
\item{args}{Argument.}

\item{verbose}{Whether to be loud}
}
\description{
Parse command line arguments and assigns the values of them. |args| is meant
to just be additional command line arguments.
}
\examples{
\dontrun{parse_args(args = commandArgs(trailingOnly = TRUE))}

}
