% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplement-package.R
\docType{package}
\name{simplement-package}
\alias{simplement}
\alias{simplement-package}
\title{simplement}
\description{
This package contains helpers for simulations and numerical experiments.
}
\author{
\strong{Maintainer}: Sangwon Hyun \email{sangwonh@ucsc.edu}

}
\keyword{internal}
